qsimd: Initialize the current state as soon as QtCore loads
Instead of lazily initializing. We usually don't do this in Qt, but in
this case there are two very good reasons for it:

1) the call tp qCpuFeatures() was not trivial, with the need to preserve
   a bit of state in the caller function across the call. GCC appeared
   to generate better code than Clang in this regard, but it still
   implied more cost than we'd like to do runtime detection in
   performance-sensitive places in Qt.

2) the early initialization allows us to use the detected state in GNU
   indirect functions on platforms that support it.

In order to do this, I had to rewrite the QT_NO_CPU_FEATURE environment
variable parsing without QByteArray and instead rely on string.h's
strtok().

This can't be done for static Qt builds on platforms that don't support
the GNU init_priority variable attribute or the MSVC equivalent[1],
because otherwise we can't guarantee that this bit of code runs before
everything else in Qt. For those platforms, we keep the existing lazy
initialization.

For shared builds in those platforms, we can use the dynamic
initialization. All known linkers will sort the static initialization
code in the order in which the .cpp are linked into the library.

The x86 QSimdInitialized variable will be removed in the next commit.

[1] https://docs.microsoft.com/en-us/cpp/preprocessor/init-seg?view=msvc-170

Change-Id: I6fcda969a9e9427198bffffd16ce885a27f6b9e2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>